requirements.txt
fastapi
uvicorn
mysql-connector-python
jinja2
python-dotenv
passlib[bcrypt]
python-multipart